Add Mandatory part

+++ Circle 00/Libft/ft_strlen.c
@@ -12,9 +12,9 @@
 #include <stdio.h>
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
